refactor: extract order processing to OrderService
