+++ api/src/main/java/com/nimbleways/springboilerplate/contollers/OrderController.java
@@ -1,0 +1,25 @@
+package com.nimbleways.springboilerplate.contollers;
+
+import com.nimbleways.springboilerplate.dto.product.ProcessOrderResponse;
+import com.nimbleways.springboilerplate.services.implementations.OrderService;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/orders")
+@RequiredArgsConstructor
+public class OrderController {
+   private final OrderService orderService;
+
+    @PostMapping("{orderId}/processOrder")
+    @ResponseStatus(HttpStatus.OK)
+    public ProcessOrderResponse processOrder(@PathVariable Long orderId) {
+        return orderService.processOrder(orderId);
+    }
+}
