+++ api/src/main/java/com/nimbleways/springboilerplate/services/implementations/OrderService.java
@@ -1,41 +1,28 @@
-package com.nimbleways.springboilerplate.contollers;
+package com.nimbleways.springboilerplate.services.implementations;
 
 import com.nimbleways.springboilerplate.dto.product.ProcessOrderResponse;
 import com.nimbleways.springboilerplate.entities.Order;
 import com.nimbleways.springboilerplate.entities.Product;
 import com.nimbleways.springboilerplate.repositories.OrderRepository;
 import com.nimbleways.springboilerplate.repositories.ProductRepository;
-import com.nimbleways.springboilerplate.services.implementations.ProductService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
 
-@RestController
-@RequestMapping("/orders")
-public class MyController {
-    @Autowired
-    private ProductService ps;
-
-    @Autowired
-    private ProductRepository pr;
-
-    @Autowired
-    private OrderRepository or;
-
-    @PostMapping("{orderId}/processOrder")
-    @ResponseStatus(HttpStatus.OK)
+@Service
+@RequiredArgsConstructor
+public class OrderService {
+    private final ProductService productService;
+    private final ProductRepository productRepository;
+    private final OrderRepository orderRepository;
     public ProcessOrderResponse processOrder(@PathVariable Long orderId) {
-        Order order = or.findById(orderId).get();
+        Order order = orderRepository.findById(orderId).get();
         System.out.println(order);
         List<Long> ids = new ArrayList<>();
         ids.add(orderId);
@@ -44,11 +31,11 @@ public class MyController {
             if (p.getType().equals("NORMAL")) {
                 if (p.getAvailable() > 0) {
                     p.setAvailable(p.getAvailable() - 1);
-                    pr.save(p);
+                    productRepository.save(p);
                 } else {
                     int leadTime = p.getLeadTime();
                     if (leadTime > 0) {
-                        ps.notifyDelay(leadTime, p);
+                        productService.notifyDelay(leadTime, p);
                     }
                 }
             } else if (p.getType().equals("SEASONAL")) {
@@ -56,19 +43,19 @@ public class MyController {
                 if ((LocalDate.now().isAfter(p.getSeasonStartDate()) && LocalDate.now().isBefore(p.getSeasonEndDate())
                         && p.getAvailable() > 0)) {
                     p.setAvailable(p.getAvailable() - 1);
-                    pr.save(p);
+                    productRepository.save(p);
                 } else {
-                    ps.handleSeasonalProduct(p);
+                    productService.handleSeasonalProduct(p);
                 }
             } else if (p.getType().equals("EXPIRABLE")) {
                 if (p.getAvailable() > 0 && p.getExpiryDate().isAfter(LocalDate.now())) {
                     p.setAvailable(p.getAvailable() - 1);
-                    pr.save(p);
+                    productRepository.save(p);
                 } else {
-                    ps.handleExpiredProduct(p);
+                    productService.handleExpiredProduct(p);
                 }
             } else if (p.getType().equals("FLASHSALE")) {
-                ps.handleFlashSaleProduct(p);
+                productService.handleFlashSaleProduct(p);
             }
         }
