+++ api/src/test/java/com/nimbleways/springboilerplate/controllers/OrderControllerIntegrationTests.java
@@ -32,7 +32,7 @@ import java.util.Set;
 @SetupDatabase
 @SpringBootTest
 @AutoConfigureMockMvc
-public class MyControllerIntegrationTests {
+public class OrderControllerIntegrationTests {
         @Autowired
         private MockMvc mockMvc;
